Fixed speedometer, first attempt at LSTM
Fixed speedometer because it was dropping to zero about every 3rd frame.
Made filemash output a sequence of low-res images along with the single
highres image. The sequence is for the LSTM. It is not yet as good as
the Alexnet. There's an assert that will break the car, but it's just
a reminder. More options in config.py to change between Alexnet and
LSTM.